P8P/requirements.txt
flask
requests
openai
pywebview
google-generativeai
beautifulsoup4
PyMuPDF
openpyxl
qrcode
uuid
